pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-trading-system"
version = "0.1.0"
description = "KOSPI200 기반 퀀트 투자 전략 백테스팅 시스템"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Quant Trading Team", email = "team@example.com"},
]
maintainers = [
    {name = "Quant Trading Team", email = "team@example.com"},
]
keywords = ["quantitative", "trading", "backtesting", "machine-learning", "finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "matplotlib>=3.8.0",
    "plotly>=5.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "finance-datareader>=0.9.0",
    "OpenDartReader>=0.2.0",
    "pykrx>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "pytest>=7.4.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "notebook>=7.0.0",
    "pyarrow>=12.0.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/username/quant-trading-system"
Documentation = "https://quant-trading-system.readthedocs.io/"
Repository = "https://github.com/username/quant-trading-system.git"
Issues = "https://github.com/username/quant-trading-system/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = [
    "baseline_*",
    ".git",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]  # assert false

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*",
    "sklearn.*",
    "xgboost.*",
    "lightgbm.*",
    "plotly.*",
    "matplotlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
